feat(vm): implement `InspectCpu` for cycle accurate CPU.
* fix(api): change state fields `ac` and `sp` type from `u16` to `u8`

+++ effnes-cpu/src/inspect.rs
@@ -2,8 +2,8 @@ use crate::{addr::AddressingMode, consts::Flags};
 
 pub struct State {
     pub pc: u16,
-    pub sp: u16,
-    pub ac: u16,
+    pub sp: u8,
+    pub ac: u8,
     pub ix: u8,
     pub iy: u8,
     pub am: AddressingMode,
